+++ lib/src/manager/state/row_group_state.dart
@@ -64,6 +64,11 @@ abstract class IRowGroupState {
   /// {@macro row_group_state_iterateRow}
   Iterable<PlutoRow> get iterateAllRow;
 
+  /// Regardless of pagination applied,
+  ///
+  /// {@macro row_group_state_iterateMainRowGroup}
+  Iterable<PlutoRow> get iterateFilteredMainRowGroup;
+
   /// With filtering or pagination applied,
   ///
   /// {@template row_group_state_iterateMainRowGroup}
@@ -191,6 +196,13 @@ mixin RowGroupState implements IPlutoGridState {
   }
 
   @override
+  Iterable<PlutoRow> get iterateFilteredMainRowGroup sync* {
+    for (final row in refRows.filterOrOriginalList.where(isMainRow)) {
+      yield row;
+    }
+  }
+
+  @override
   Iterable<PlutoRow> get iterateMainRowGroup sync* {
     for (final row in refRows.where(isMainRow)) {
       yield row;
@@ -283,6 +295,13 @@ mixin RowGroupState implements IPlutoGridState {
 
     if (isPaginated) {
       resetPage(resetCurrentState: false, notify: false);
+    }
+
+    if (rowGroupDelegate?.onToggled != null) {
+      rowGroupDelegate!.onToggled!(
+        row: rowGroup,
+        expanded: rowGroup.type.group.expanded,
+      );
     }
 
     updateCurrentCellPosition(notify: false);