+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [5.4.3] - 2022. 11. 21
+
+* Add iterateRowType to PlutoAggregateColumnFooter.
+* Modify enter key.(Included numpadEnter)
+* Add onToggled to row group delegate.
+
 ## [5.4.2] - 2022. 11. 15
 
 * Add properties of PlutoGridScrollbarConfig.